+++ packages/block-editor/src/components/inserter/quick-inserter.js
@@ -1,4 +1,10 @@
 /**
+ * External dependencies
+ */
+import { orderBy } from 'lodash';
+import classnames from 'classnames';
+
+/**
  * WordPress dependencies
  */
 import { useState, useMemo, useEffect } from '@wordpress/element';
@@ -8,7 +14,7 @@ import {
 	Button,
 	withSpokenMessages,
 } from '@wordpress/components';
-import { useSelect } from '@wordpress/data';
+import { useSelect, useDispatch } from '@wordpress/data';
 import { LEFT, RIGHT, UP, DOWN, BACKSPACE, ENTER } from '@wordpress/keycodes';
 
 /**
@@ -46,7 +52,11 @@ function QuickInserterList( {
 	onHover,
 } ) {
 	const shownBlockTypes = useMemo(
-		() => blockTypes.slice( 0, SHOWN_BLOCK_TYPES ),
+		() =>
+			orderBy( blockTypes, [ 'frecency' ], [ 'desc' ] ).slice(
+				0,
+				SHOWN_BLOCK_TYPES
+			),
 		[ blockTypes ]
 	);
 	const shownBlockPatterns = useMemo(
@@ -69,6 +79,8 @@ function QuickInserterList( {
 						items={ shownBlockTypes }
 						onSelect={ onSelectBlockType }
 						onHover={ onHover }
+						label={ __( 'Blocks' ) }
+						limit={ SHOWN_BLOCK_TYPES }
 					/>
 				</InserterPanel>
 			) }
@@ -97,6 +109,7 @@ function QuickInserterList( {
 }
 
 function QuickInserter( {
+	onSelect,
 	rootClientId,
 	clientId,
 	isAppender,
@@ -109,6 +122,7 @@ function QuickInserter( {
 		onInsertBlocks,
 		onToggleInsertionPoint,
 	] = useInsertionPoint( {
+		onSelect,
 		rootClientId,
 		clientId,
 		isAppender,
@@ -143,18 +157,26 @@ function QuickInserter( {
 		[ filterValue, patterns ]
 	);
 
-	const setInsererIsOpened = useSelect(
+	const setInserterIsOpened = useSelect(
 		( select ) =>
 			select( 'core/block-editor' ).getSettings()
 				.__experimentalSetIsInserterOpened,
 		[]
 	);
 
+	const previousBlockClientId = useSelect(
+		( select ) =>
+			select( 'core/block-editor' ).getPreviousBlockClientId( clientId ),
+		[ clientId ]
+	);
+
 	useEffect( () => {
-		if ( setInsererIsOpened ) {
-			setInsererIsOpened( false );
+		if ( setInserterIsOpened ) {
+			setInserterIsOpened( false );
 		}
-	}, [ setInsererIsOpened ] );
+	}, [ setInserterIsOpened ] );
+
+	const { selectBlock } = useDispatch( 'core/block-editor' );
 
 	// Announce search results on change
 	useEffect( () => {
@@ -170,6 +192,19 @@ function QuickInserter( {
 		debouncedSpeak( resultsFoundMessage );
 	}, [ filterValue, debouncedSpeak ] );
 
+	// When clicking Browse All select the appropriate block so as
+	// the insertion point can work as expected
+	const onBrowseAll = () => {
+		// We have to select the previous block because the menu inserter
+		// inserts the new block after the selected one.
+		// Ideally, this selection shouldn't focus the block to avoid the setTimeout.
+		selectBlock( previousBlockClientId );
+		// eslint-disable-next-line @wordpress/react-no-unsafe-timeout
+		setTimeout( () => {
+			setInserterIsOpened( true );
+		} );
+	};
+
 	// Disable reason (no-autofocus): The inserter menu is a modal display, not one which
 	// is always visible, and one which already incurs this behavior of autoFocus via
 	// Popover's focusOnMount.
@@ -178,7 +213,10 @@ function QuickInserter( {
 	/* eslint-disable jsx-a11y/no-autofocus, jsx-a11y/no-static-element-interactions */
 	return (
 		<div
-			className="block-editor-inserter__quick-inserter"
+			className={ classnames( 'block-editor-inserter__quick-inserter', {
+				'has-search': showSearch,
+				'has-expand': setInserterIsOpened,
+			} ) }
 			onKeyPress={ stopKeyPropagation }
 			onKeyDown={ preventArrowKeysPropagation }
 		>
@@ -199,10 +237,10 @@ function QuickInserter( {
 				onHover={ onToggleInsertionPoint }
 			/>
 
-			{ setInsererIsOpened && (
+			{ setInserterIsOpened && (
 				<Button
 					className="block-editor-inserter__quick-inserter-expand"
-					onClick={ () => setInsererIsOpened( true ) }
+					onClick={ onBrowseAll }
 					aria-label={ __(
 						'Browse all. This will open the main inserter panel in the editor toolbar.'
 					) }