+++ packages/block-editor/src/components/keyboard-shortcuts/index.js
@@ -131,7 +131,7 @@ function KeyboardShortcuts() {
 			},
 			[ clientIds, removeBlocks ]
 		),
-		{ isDisabled: clientIds.length < 1 }
+		{ isDisabled: clientIds.length < 2 }
 	);
 
 	useShortcut(