+++ packages/block-editor/src/components/preview-options/index.js
@@ -6,7 +6,7 @@ import classnames from 'classnames';
 /**
  * WordPress dependencies
  */
-import { Button, Dropdown, MenuGroup, MenuItem } from '@wordpress/components';
+import { DropdownMenu, MenuGroup, MenuItem } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
 import { check } from '@wordpress/icons';
 
@@ -17,27 +17,27 @@ export default function PreviewOptions( {
 	deviceType,
 	setDeviceType,
 } ) {
+	const popoverProps = {
+		className: classnames(
+			className,
+			'block-editor-post-preview__dropdown-content'
+		),
+		position: 'bottom left',
+	};
+	const toggleProps = {
+		isTertiary: true,
+		className: 'block-editor-post-preview__button-toggle',
+		disabled: ! isEnabled,
+		children: __( 'Preview' ),
+	};
 	return (
-		<Dropdown
+		<DropdownMenu
 			className="block-editor-post-preview__dropdown"
-			contentClassName={ classnames(
-				className,
-				'block-editor-post-preview__dropdown-content'
-			) }
-			popoverProps={ { role: 'menu' } }
-			position="bottom left"
-			renderToggle={ ( { isOpen, onToggle } ) => (
-				<Button
-					isTertiary
-					onClick={ onToggle }
-					className="block-editor-post-preview__button-toggle"
-					aria-expanded={ isOpen }
-					disabled={ ! isEnabled }
-				>
-					{ __( 'Preview' ) }
-				</Button>
-			) }
-			renderContent={ () => (
+			popoverProps={ popoverProps }
+			toggleProps={ toggleProps }
+			icon={ null }
+		>
+			{ () => (
 				<>
 					<MenuGroup>
 						<MenuItem
@@ -65,6 +65,6 @@ export default function PreviewOptions( {
 					{ children }
 				</>
 			) }
-		/>
+		</DropdownMenu>
 	);
 }