+++ packages/block-library/src/index.js
@@ -65,6 +65,8 @@ import * as socialLink from './social-link';
 import * as widgetArea from './widget-area';
 
 // Full Site Editing Blocks
+import * as siteLogo from './site-logo';
+import * as siteTagline from './site-tagline';
 import * as siteTitle from './site-title';
 import * as templatePart from './template-part';
 import * as query from './query';
@@ -132,8 +134,6 @@ export const registerCoreBlocks = () => {
 		column,
 		cover,
 		embed,
-		...embed.common,
-		...embed.others,
 		file,
 		group,
 		window.wp && window.wp.oldEditor ? classic : null, // Only add the classic block in WP Context
@@ -198,6 +198,8 @@ export const __experimentalRegisterExperimentalCoreBlocks =
 					// Register Full Site Editing Blocks.
 					...( __experimentalEnableFullSiteEditing
 						? [
+								siteLogo,
+								siteTagline,
 								siteTitle,
 								templatePart,
 								query,