+++ packages/block-library/src/template-part/edit/name-panel.js
@@ -28,8 +28,9 @@ export default function TemplatePartNamePanel( { postId, setAttributes } ) {
 					setTitle( value );
 					const newSlug = cleanForSlug( value );
 					setSlug( newSlug );
-					setAttributes( { slug: newSlug } );
+					setAttributes( { slug: newSlug, postId } );
 				} }
+				onFocus={ ( event ) => event.target.select() }
 			/>
 		</div>
 	);