+++ packages/e2e-tests/config/setup-test-framework.js
@@ -34,7 +34,14 @@ const THROTTLE_CPU = process.env.THROTTLE_CPU;
  *
  * @type {string|undefined}
  */
-const DOWNLOAD_THROUGHPUT = process.env.DOWNLOAD_THROUGHPUT;
+const SLOW_NETWORK = process.env.SLOW_NETWORK;
+
+/**
+ * Emulate no internet connection.
+ *
+ * @type {string|undefined}
+ */
+const OFFLINE = process.env.OFFLINE;
 
 /**
  * Set of console logging types observed to protect against unexpected yet
@@ -137,14 +144,6 @@ function observeConsoleLogging() {
 			return;
 		}
 
-		// As of WordPress 5.3.2 in Chrome 79, navigating to the block editor
-		// (Posts > Add New) will display a console warning about
-		// non - unique IDs.
-		// See: https://core.trac.wordpress.org/ticket/23165
-		if ( text.includes( 'elements with non-unique id #_wpnonce' ) ) {
-			return;
-		}
-
 		const logFunction = OBSERVED_CONSOLE_MESSAGE_TYPES[ type ];
 
 		// As of Puppeteer 1.6.1, `message.text()` wrongly returns an object of
@@ -221,17 +220,24 @@ async function runAxeTestsForBlockEditor() {
  * Simulate slow network or throttled CPU if provided via environment variables.
  */
 async function simulateAdverseConditions() {
-	if ( ! DOWNLOAD_THROUGHPUT && ! THROTTLE_CPU ) {
+	if ( ! SLOW_NETWORK && ! OFFLINE && ! THROTTLE_CPU ) {
 		return;
 	}
 
 	const client = await page.target().createCDPSession();
 
-	if ( DOWNLOAD_THROUGHPUT ) {
+	if ( SLOW_NETWORK || OFFLINE ) {
 		// See: https://chromedevtools.github.io/devtools-protocol/tot/Network#method-emulateNetworkConditions
+		// The values below simulate fast 3G conditions as per https://github.com/ChromeDevTools/devtools-frontend/blob/80c102878fd97a7a696572054007d40560dcdd21/front_end/sdk/NetworkManager.js#L252-L274
 		await client.send( 'Network.emulateNetworkConditions', {
-			// Simulated download speed (bytes/s)
-			downloadThroughput: Number( DOWNLOAD_THROUGHPUT ),
+			// Network connectivity is absent
+			offline: Boolean( OFFLINE || false ),
+			// Download speed (bytes/s)
+			downloadThroughput: ( ( 1.6 * 1024 * 1024 ) / 8 ) * 0.9,
+			// Upload speed (bytes/s)
+			uploadThroughput: ( ( 750 * 1024 ) / 8 ) * 0.9,
+			// Latency (ms)
+			latency: 150 * 3.75,
 		} );
 	}